New (imo improved) artist parsing
Main functionality is removing regex and relying on native functions. Beyond that, is a fair bit more rigorous (for instance, if there was an artist with a dash in their name, say blume - popo, "blume - popo -- title" would be interpreted as blume, not blume - popo). Does take up more space though.