New (imo improved) artist parsing (#33)
* New (imo improved) artist parsing

Main functionality is removing regex and relying on native functions. Beyond that, is a fair bit more rigorous (for instance, if there was an artist with a dash in their name, say blume - popo, "blume - popo -- title" would be interpreted as blume, not blume - popo). Does take up more space though.

* Formatting mostly

Fixed formatting and naming as requested by martijnboers. Also added an @staticmethod because I forgot to add it in initial commit.

* changed double quotes to single quotes

* Inlined a variable that was only used once

* extracted title parsing to separate file

* Fixed missing featuring condition

* Updated test

* Fixed undefined variables

* Add files via upload